Add files via upload
This project is somehow "unorthodox", normally betting company rate the expected results based on different criteria and usually they used a structured way to do so. But there is some "rumor" that the match results are influenced by the interest of betting companies. In this work, I try to explore, is there any "reverse" relationship, so we took the data from different betting websites  and  based on those try to predict the match results and see how accurate they are.